tried to fix leak and norminette
